Resolve RKE Template issue caused by CIS
The profile helper methods were attached to the cluster model.
Unfortunately, the cluster isn't available when creating a new rke
template.

To resolve this I moved all of the cis helpers out of the cluster model
and utils and moved them into a cisHelpers service so they could be
used without access to the cluster itself.
