+++ app/models/cluster.js
@@ -9,7 +9,6 @@ import { equal, alias } from '@ember/object/computed';
 import { resolve } from 'rsvp';
 import C from 'ui/utils/constants';
 import { isEmpty } from '@ember/utils';
-import { createProfileKey, toTitle } from 'shared/utils/util';
 import moment from 'moment';
 const TRUE = 'True';
 const CLUSTER_TEMPLATE_ID_PREFIX = 'cattle-global-data:';
@@ -411,45 +410,6 @@ export default Resource.extend(Grafana, ResourceUsage, {
     });
 
     return out;
-  }),
-
-
-  cisScanConfigProfiles: computed(function() {
-    return this.globalStore.getById('schema', 'cisscanconfig').optionsFor('profile');
-  }),
-
-  cisScanBenchmarks: computed(() => {
-    return [
-      'rke-cis-1.4',
-      'rke-cis-1.5'
-    ]
-  }),
-
-  cisScanProfiles: computed('cisScanConfigProfiles', 'cisScanBenchmarks', function() {
-    const profiles = get(this, 'cisScanConfigProfiles');
-    const benchmarks = get(this, 'cisScanBenchmarks');
-
-    const asArray = profiles.flatMap((profile) => {
-      return benchmarks.map((benchmark) => ({
-        [createProfileKey(profile, benchmark)]: {
-          benchmark,
-          profile
-        }
-      }))
-    });
-
-    return Object.assign.apply({}, asArray);
-  }),
-
-  cisScanProfileOptions: computed('cisScanProfiles', function() {
-    return Object.keys(get(this, 'cisScanProfiles')).map((key) => ({
-      label: toTitle(key),
-      value: key
-    }))
-  }),
-
-  defaultCisScanProfileOption: computed('cisScanProfileOptions', function() {
-    return get(this, 'cisScanProfileOptions')[0].value;
   }),
 
   actions: {
