+++ lib/shared/addon/components/cluster-driver/driver-rke/component.js
@@ -5,9 +5,7 @@ import {
 } from '@ember/object';
 import { maxSatisfying } from 'shared/utils/parse-version';
 import { inject as service } from '@ember/service';
-import {
-  clusterScanConfigToProfile, removeEmpty, keysToCamel, profileToClusterScanConfig, validateEndpoint, keysToDecamelize,
-} from 'shared/utils/util';
+import { removeEmpty, keysToCamel, validateEndpoint, keysToDecamelize } from 'shared/utils/util';
 import { validateHostname } from '@rancher/ember-api-store/utils/validate';
 import { validateCertWeakly } from 'shared/utils/util';
 import C from 'shared/utils/constants';
@@ -79,6 +77,7 @@ export default InputTextFile.extend(ManageLabels, ClusterDriver, {
   clusterTemplates:             service(),
   access:                       service(),
   router:                       service(),
+  cisHelpers:                   service(),
 
   layout,
   authChoices:                  AUTHCHOICES,
@@ -929,10 +928,9 @@ export default InputTextFile.extend(ManageLabels, ClusterDriver, {
 
   initScheduledClusterScan() {
     const scheduledClusterScan = get(this, 'primaryResource.scheduledClusterScan') || get(this, 'scheduledClusterScan');
-    const scanConfig = get(this, 'primaryResource.scheduledClusterScan.scanConfig') || profileToClusterScanConfig(get(this, 'model.cluster.defaultCisScanProfileOption'), get(this, 'model.cluster.cisScanProfiles'));
-    const scheduleConfig = get(this, 'primaryResource.scheduledClusterScan.scheduleConfig') || get(this, 'scheduledClusterScan.scheduleConfig');
-    const cisProfile = clusterScanConfigToProfile(scanConfig);
-
+    const scanConfig = get(this, 'primaryResource.scheduledClusterScan.scanConfig') || this.cisHelpers.defaultClusterScanConfig;
+    const scheduleConfig = get(this, 'primaryResource.scheduledClusterScan.scheduleConfig') ||  get(this, 'scheduledClusterScan.scheduleConfig');
+    const cisProfile =  this.cisHelpers.clusterScanConfigToProfile(scanConfig);
 
     set(this, 'scheduledClusterScan', scheduledClusterScan);
     set(this, 'scheduledClusterScan.scanConfig', scanConfig);
@@ -1049,7 +1047,7 @@ export default InputTextFile.extend(ManageLabels, ClusterDriver, {
     const scheduledClusterScan = { ...get(this, 'scheduledClusterScan') };
 
     if (scheduledClusterScan.enabled) {
-      scheduledClusterScan.scanConfig = profileToClusterScanConfig(get(this, 'cisProfile'), get(this, 'model.cluster.cisScanProfiles'));
+      scheduledClusterScan.scanConfig = this.cisHelpers.profileToClusterScanConfig(get(this, 'cisProfile'));
     } else {
       scheduledClusterScan.scanConfig = null;
       scheduledClusterScan.scheduleConfig = null;
