+++ lib/shared/addon/components/cluster-driver/driver-rke/template.hbs
@@ -994,7 +994,7 @@
                   {{new-select
                     id="cis-scan-profile"
                     classNames="form-control"
-                    content=model.cluster.cisScanProfileOptions
+                    content=cisHelpers.cisScanProfileOptions
                     value=cisProfile
                     disabled=(not scheduledClusterScan.enabled)
                   }}
