+++ lib/shared/addon/components/run-scan-modal/component.js
@@ -9,6 +9,7 @@ export default Component.extend(ModalBase, {
   globalStore:  service(),
   growl:        service(),
   intl:         service(),
+  cisHelpers:   service(),
   modalService: service('modal'),
 
   layout,
@@ -18,8 +19,7 @@ export default Component.extend(ModalBase, {
 
   init() {
     this._super(...arguments);
-
-    set(this, 'profile', get(this, 'modalOpts.cluster.cisScanProfileOptions.first.value'));
+    set(this, 'profile', this.cisHelpers.cisScanProfileOptions[0].value);
   },
 
   actions: {
@@ -28,7 +28,9 @@ export default Component.extend(ModalBase, {
       const onRun = get(this, 'modalOpts.onRun');
       const intl = get(this, 'intl');
 
-      const profile = get(this, 'modalOpts.cluster.cisScanProfiles')[get(this, 'profile')];
+      const profile = this.cisHelpers.cisScanProfiles[get(this, 'profile')];
+      const clusterName = get(this, 'modalOpts.cluster.displayName');
+
 
       cluster.doAction('runSecurityScan', {
         failuresOnly:             false,
@@ -36,7 +38,7 @@ export default Component.extend(ModalBase, {
         profile:                  profile.profile,
         overrideBenchmarkVersion: profile.benchmark
       }).then(() => {
-        this.growl.success(intl.t('cis.scan.growl.success', { clusterName: get(this, 'name') }), '');
+        this.growl.success(intl.t('cis.scan.growl.success', { clusterName }), '');
       });
 
       this.get('modalService').toggleModal();
