+++ lib/shared/addon/components/run-scan-modal/template.hbs
@@ -8,7 +8,7 @@
 {{new-select
   id="cis-scan-profile"
   classNames="form-control"
-  content=modalOpts.cluster.cisScanProfileOptions
+  content=cisHelpers.cisScanProfileOptions
   value=profile
 }}
