+++ lib/shared/addon/utils/util.js
@@ -582,29 +582,6 @@ export function extractUniqueStrings(strings) {
   return Object.keys(index);
 }
 
-export function createProfileKey(profile, benchmark) {
-  return `${ benchmark.toUpperCase() } ${ profile } `;
-}
-
-export function clusterScanConfigToProfile(scanConfig) {
-  console.log(scanConfig.cisScanConfig.profile, scanConfig.cisScanConfig.overrideBenchmarkVersion);
-
-  return createProfileKey(scanConfig.cisScanConfig.profile, scanConfig.cisScanConfig.overrideBenchmarkVersion);
-}
-
-export function profileToClusterScanConfig(profile, cisScanProfiles) {
-  const profileBenchmark = cisScanProfiles[profile];
-
-  return {
-    cisScanConfig: {
-      failuresOnly:             false,
-      skip:                     null,
-      profile:                  profileBenchmark.profile,
-      overrideBenchmarkVersion: profileBenchmark.benchmark
-    }
-  }
-}
-
 var Util = {
   absoluteUrl,
   addAuthorization,
@@ -614,10 +591,8 @@ var Util = {
   arrayIntersect,
   compareDisplayEndpoint,
   camelToTitle,
-  clusterScanConfigToProfile,
   convertToMillis,
   constructUrl,
-  createProfileKey,
   download,
   deepCopy,
   escapeHtml,
@@ -638,7 +613,6 @@ var Util = {
   parseUrl,
   pluralize,
   popupWindowOptions,
-  profileToClusterScanConfig,
   random32,
   randomStr,
   removeEmpty,
