+++ lib/shared/app/cis-helpers/service.js
@@ -1,0 +1,1 @@
+export { default } from 'shared/cis-helpers/service';
